pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-learning-2026"
version = "0.1.0"
description = "A systematic RAG learning project for 2026"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "RAG Learning Team"}
]
keywords = ["rag", "retrieval", "llm", "nlp", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "sentence-transformers>=2.2.0",
    "langchain>=0.1.0",
    "llama-index>=0.9.0",
    "faiss-cpu>=1.7.4",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.7.0",
]
all = [
    "openai>=1.0.0",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
]

[project.urls]
Homepage = "https://github.com/your-username/RAG-Learning-2026"
Documentation = "https://rag-learning.readthedocs.io"
Repository = "https://github.com/your-username/RAG-Learning-2026.git"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
